Detached WelcomeView from WelcomeViewController
