+++ WeatherApp/AppDelegate.swift
@@ -1,10 +1,3 @@
-//
-//  AppDelegate.swift
-//  WeatherApp
-//
-//  Created by Oskar Biwejnis on 11/07/2023.
-//
-
 import UIKit
 @main
 class AppDelegate: UIResponder, UIApplicationDelegate {
