+++ WeatherApp/SceneDelegate.swift
@@ -1,10 +1,3 @@
-//
-//  SceneDelegate.swift
-//  WeatherApp
-//
-//  Created by Oskar Biwejnis on 11/07/2023.
-//
-
 import UIKit
 
 class SceneDelegate: UIResponder, UIWindowSceneDelegate {
